fix TodoAddInput component
